refactor: Rename Home component
